Fix a lot of things
Again, we're fixing more things than we should. There's definitely some
silly stuff going on that should be cleaned up eventually. But, at least
it's coming along. Logos are highlighting the way they should now and
solved a tricky bug that only manifested if you'd highlight the text and
instead of deleting it, you'd type a number in, replacing the text with
just that one character. It wasn't setting the previous flag for logo
visibility to false again, so I forced it.