Grab all prefixes and place in separate file
Initially, the prefix list was not complete as it is a large list. It
was used for testing at the start when the list didn't need to be
complete. However, now the full list is required to allow for all of the
numbers to be tested and utilized for verification.

The list was separated out into its own file because of its size. It
would unnecessarily pollute `App.js` otherwise.